feat: label speakers in the gutter instead of re-cutting the transcript
